[UBA] * Fixed linux compile error
[CL 33916910 by henrik karlsson in ue5-main branch]

+++ Engine/Source/Programs/UnrealBuildAccelerator/Common/Private/UbaCacheServer.cpp
@@ -1192,7 +1192,7 @@ namespace uba
 					{
 						CacheEntries& entries = kv2.second;
 						SCOPED_READ_LOCK(entries.lock, lock3);
-						mostEntries = Max(mostEntries, entries.entries.size());
+						mostEntries = Max(mostEntries, u64(entries.entries.size()));
 						for (auto& entry : entries.entries)
 							lastUsed = Max(lastUsed, entry.lastUsedTime);
 					}
